remove direct dependency on sslr-squid-bridge and guava
these libraries are used by sonar-java, so sonar-findbugs can use them
without depending on them directly. And in SonarQube, libraries in sonar-java
are used because useChildFirstClassLoader is false for sonar-findbugs.

so we can simply remove these libraries from our pom file.

sonar-testing-harness is not used so we remove it in this commit too.