Bundle the dependencies in the jar lib folder instead of shading them
When building the dependencies programmatically in the sq-10 branch we
need to have a findbugs.xml file per plugin on the classpath. The maven
shade plugin only keeps one, so only the core plugin loads.
Instead of shading the dependencies we can let the sonar packaging
plugin bundle them in the lib folder of the jar.
Since the sonar-java-plugin already shades its own dependencies we do
not need to bundle its transient dependencies, this makes the final jar
significantly smaller in particular because the ECJ jar is quite large
at around 10Mo.