Fix end of file markers
git will add end of file markers for its usage so this ensures files are accurately in repo so that cost is not incurred, if using github you will see it flagged regarding this and this fixes the issue.  Its a one time effort typically unless someone introduces incorrect file endings again.